Added pressure difference
Code is not working correctly

+++ NS/visual.h
@@ -31,7 +31,7 @@ void write_vtkFile(const char *szProblem,
 		  double dy,
                   double **U,
                   double **V,
-                  double **P);
+                  int **P);
 
 /**
  * Method for writing header information in vtk format.
